+++ tests/termination.rs
@@ -101,7 +101,7 @@ fn test_tol_fun() {
     let function = |x: &DVector<f64>| 1.0 + x.magnitude().powi(2);
     run_test(
         function,
-        CMAESOptions::new(2).initial_mean(vec![5.0; 2]),
+        CMAESOptions::new(2).tol_fun_hist(0.0).initial_mean(vec![5.0; 2]),
         |r| matches!(r, TerminationReason::TolFun),
         0,
     );
@@ -120,13 +120,13 @@ fn test_tol_x() {
 }
 
 #[test]
-fn test_equal_fun_values() {
+fn test_tol_fun_hist() {
     // The function bottoms out before convergence
     let function = |x: &DVector<f64>| x.magnitude().max(1e-6);
     run_test(
         function,
         CMAESOptions::new(2).tol_fun(0.0).initial_mean(vec![5.0; 2]),
-        |r| matches!(r, TerminationReason::EqualFunValues),
+        |r| matches!(r, TerminationReason::TolFunHist),
         0,
     );
 }
@@ -163,7 +163,7 @@ fn run_test_no_effect<F: Fn(TerminationReason) -> bool + Clone>(check_reason: F)
         |x: &DVector<f64>| 1e-8 + (2.0 * x[0] - x[1]).abs().powf(1.5) + (2.0 - x[1]).powi(2);
     run_test(
         function,
-        CMAESOptions::new(2).tol_fun(0.0).tol_x(1e-16).initial_step_size(4.0),
+        CMAESOptions::new(2).tol_fun(0.0).tol_fun_hist(0.0).tol_x(1e-16).initial_step_size(4.0),
         check_reason,
         1,
     );